requirements.txt
absl-py==2.1.0
aioice==0.9.0
aiortc==1.9.0
altair==4.2.0
astunparse==1.6.3
attrs==23.2.0
av==12.1.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cryptography==42.0.8
cycler==0.12.1
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.53.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.30.0
google-auth-oauthlib==1.0.0
google-crc32c==1.5.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
ifaddr==0.2.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==2.13.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pandas==2.0.3
pillow==10.3.0
pkgutil_resolve_name==1.3.10
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydeck==0.9.1
pyee==11.1.0
Pygments==2.18.0
pylibsrtp==0.10.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyshine==0.0.9
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
six==1.16.0
smmap==5.0.1
sounddevice==0.4.7
streamlit==1.35.0
streamlit-webrtc==0.47.7
tenacity==8.4.1
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.5.0
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
